requirements.txt
streamlit==1.30.0
streamlit-folium==0.25.1
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
pydeck==0.8.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
folium==0.14.0
branca==0.6.0
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
xlrd==2.0.1
geopy==2.4.1
python-dotenv==1.0.0
streamlit-aggrid